Feature/1456 Implement zone arrays for soil parameters
* Implement parameter input fields to ZoneEditor
* Rename style class inputfix to inputFix
