+++ src/t03/components/optimization/optimizationConstraints.jsx
@@ -101,7 +101,7 @@ class OptimizationConstraintsComponent extends React.Component {
                 width: 'auto',
                 height: 'auto'
             },
-            inputfix: {
+            inputFix: {
                 padding: '0'
             },
             link: {
