+++ src/t03/components/optimization/optimizationObjectives.jsx
@@ -101,7 +101,7 @@ class OptimizationObjectivesComponent extends React.Component {
                 width: 'auto',
                 height: 'auto'
             },
-            inputfix: {
+            inputFix: {
                 padding: '0'
             },
             link: {
