+++ src/t03/components/optimization/optimizationParameters.jsx
@@ -48,7 +48,7 @@ class OptimizationParametersComponent extends React.Component {
                 width: 'auto',
                 height: 'auto'
             },
-            inputfix: {
+            inputFix: {
                 padding: '0'
             },
             link: {
